+++ crates/algokit_transact_ffi/src/transactions/mod.rs
@@ -1,8 +1,10 @@
 mod application_call;
 pub mod asset_config;
+pub mod asset_freeze;
 
 pub use application_call::ApplicationCallTransactionFields;
 pub use asset_config::AssetConfigTransactionFields;
 
 pub mod keyreg;
+pub use asset_freeze::AssetFreezeTransactionFields;
 pub use keyreg::KeyRegistrationTransactionFields;
